boxOffice를 나타내는 Cell과 LiveContent를 나타내는 Cell의 분리. (둘 다 공통으로 HomeRankingContentCell 을 상속함)
